Refactor Rotator into modular package structure
- Created apts/opticalequipment/rotator/ package
- Moved base class to base.py
- Distributed 50 vendor specs and factory methods into 20 modules in vendors/
- Implemented dynamic aggregation in __init__.py to maintain public API
- Removed original rotator.py
