+++ apts/opticalequipment/rotator/vendors/altair.py
@@ -1,0 +1,40 @@
+from ..base import Rotator
+
+
+class AltairRotator(Rotator):
+    _DATABASE = {
+        "Altair_Field_Rotator_M42": {
+            "brand": "Altair",
+            "name": "Field Rotator (M42)",
+            "type": "type_rotator",
+            "optical_length": 10,
+            "mass": 240,
+            "tside_thread": "M42",
+            "tside_gender": "Male",
+            "cside_thread": "M42",
+            "cside_gender": "Female",
+            "reversible": False,
+            "bf_role": "",
+        },
+        "Altair_Field_Rotator_M48": {
+            "brand": "Altair",
+            "name": "Field Rotator (M48)",
+            "type": "type_rotator",
+            "optical_length": 11,
+            "mass": 280,
+            "tside_thread": "M48",
+            "tside_gender": "Male",
+            "cside_thread": "M48",
+            "cside_gender": "Female",
+            "reversible": False,
+            "bf_role": "",
+        },
+    }
+
+    @classmethod
+    def Altair_Field_Rotator_M42(cls):
+        return cls.from_database(cls._DATABASE["Altair_Field_Rotator_M42"])
+
+    @classmethod
+    def Altair_Field_Rotator_M48(cls):
+        return cls.from_database(cls._DATABASE["Altair_Field_Rotator_M48"])
